fix: `qfixPane` wasn't properly set when `jump` command was performed to go back
Setting the active pane when jumping back to `qfixPane` was hardcoded to index 1.
This obviously will not work correctly if `qfixPane` is not the #1. To set it to
`qfixPane` use true, instead of false, in the —already used— 'SetActive' method.

Global var `active` has been renamed to `qfixPaneTabIdx`, a more meaningful name.
Global var `tab` was not necessary, so has been removed.