fix: use "Tab.SetActive()" to activate `qfix.pane` when returning to it.
When the qfix entry was jumping within the same tab (with both the qfix pane and
the destination buffer in the same tab), the active pane wasn't reset properly
upon returning. The cursor would jump back to the qfix pane, but the line that
we could move remained in the buffer we had jumped to.

By using "Tab.SetActive()", we ensure that only the qfix pane is active when
returning.